Refactor all flag-related calls into own module and remove bus from API docs
