+++ charms/reactive/bus.py
@@ -24,8 +24,6 @@ from functools import partial
 
 from charmhelpers.core import hookenv
 from charmhelpers.core import unitdata
-from charmhelpers.cli import cmdline
-from charms.reactive import deprecated
 
 
 _log_opts = os.environ.get('REACTIVE_LOG_OPTS', '').split(',')
@@ -40,158 +38,6 @@ class BrokenHandlerException(Exception):
                    "execution failed. Only handler files may be marked "
                    "as executable.".format(path))
         super(BrokenHandlerException, self).__init__(message)
-
-
-class State(str):
-    """
-    DEPRECATED
-
-    A reactive state that can be set.
-
-    States are essentially just strings, but this class should be used to enable them
-    to be discovered and introspected, for documentation, composition, or linting.
-
-    This should be used with :class:`StateList`.
-    """
-    pass
-
-
-class StateList(object):
-    """
-    DEPRECATED
-
-    Base class for a set of states that can be set by a relation or layer.
-
-    This class should be used so that they can be discovered and introspected,
-    for documentation, composition, or linting.
-
-    Example usage::
-
-        class MyRelation(RelationBase):
-            class states(StateList):
-                connected = State('{relation_name}.connected')
-                available = State('{relation_name}.available')
-    """
-    pass
-
-
-@cmdline.subcommand()
-@cmdline.no_output
-def set_flag(flag, value=None):
-    """
-    Set the given flag as active.
-
-    :param str flag: Name of flag to set.
-    :param value: For internal use only.
-    """
-    old_flags = get_flags()
-    unitdata.kv().update({flag: value}, prefix='reactive.states.')
-    if flag not in old_flags:
-        FlagWatch.change(flag)
-
-
-@cmdline.subcommand()
-@cmdline.no_output
-def set_state(state, value=None):
-    """DEPRECATED Alias of set_flag"""
-    set_flag(state, value)
-
-
-@cmdline.subcommand()
-@cmdline.no_output
-def clear_flag(flag):
-    """
-    Clear / deactivate a flag.
-    """
-    old_flags = get_flags()
-    unitdata.kv().unset('reactive.states.%s' % flag)
-    unitdata.kv().set('reactive.dispatch.removed_state', True)
-    if flag in old_flags:
-        FlagWatch.change(flag)
-
-
-@cmdline.subcommand()
-@cmdline.no_output
-def remove_state(state):
-    """DEPRECATED Alias of clear_flag"""
-    clear_flag(state)
-
-
-@cmdline.subcommand()
-def get_flags():
-    """
-    Return a list of all flags which are set.
-    """
-    flags = unitdata.kv().getrange('reactive.states.', strip=True) or {}
-    return flags.keys()
-
-
-@cmdline.subcommand()
-def get_states():
-    """
-    DEPRECATED Use `get_flags` instead.
-
-    Return a mapping of all active states to their values.
-    """
-    return unitdata.kv().getrange('reactive.states.', strip=True) or {}
-
-
-class FlagWatch(object):
-    key = 'reactive.state_watch'
-
-    @classmethod
-    def _store(cls):
-        return unitdata.kv()
-
-    @classmethod
-    def _get(cls):
-        return cls._store().get(cls.key, {
-            'iteration': 0,
-            'changes': [],
-            'pending': [],
-        })
-
-    @classmethod
-    def _set(cls, data):
-        cls._store().set(cls.key, data)
-
-    @classmethod
-    def reset(cls):
-        cls._store().unset(cls.key)
-
-    @classmethod
-    def iteration(cls, i):
-        data = cls._get()
-        data['iteration'] = i
-        cls._set(data)
-
-    @classmethod
-    def watch(cls, watcher, flags):
-        data = cls._get()
-        iteration = data['iteration']
-        changed = bool(set(flags) & set(data['changes']))
-        return iteration == 0 or changed
-
-    @classmethod
-    def change(cls, flag):
-        data = cls._get()
-        data['pending'].append(flag)
-        cls._set(data)
-
-    @classmethod
-    def commit(cls):
-        data = cls._get()
-        data['changes'] = data['pending']
-        data['pending'] = []
-        cls._set(data)
-
-
-@deprecated.alias('get_state')
-def get_flag_value(flag, default=None):
-    """
-    For internal use only.
-    """
-    return unitdata.kv().get('reactive.states.%s' % flag, default)
 
 
 def _action_id(action):
@@ -394,6 +240,56 @@ class ExternalHandler(Handler):
         # are, and write flags (flush releases lock)
         unitdata.kv().flush()
         subprocess.check_call([self._filepath, '--invoke', self._test_output], env=os.environ)
+
+
+class FlagWatch(object):
+    key = 'reactive.state_watch'
+
+    @classmethod
+    def _store(cls):
+        return unitdata.kv()
+
+    @classmethod
+    def _get(cls):
+        return cls._store().get(cls.key, {
+            'iteration': 0,
+            'changes': [],
+            'pending': [],
+        })
+
+    @classmethod
+    def _set(cls, data):
+        cls._store().set(cls.key, data)
+
+    @classmethod
+    def reset(cls):
+        cls._store().unset(cls.key)
+
+    @classmethod
+    def iteration(cls, i):
+        data = cls._get()
+        data['iteration'] = i
+        cls._set(data)
+
+    @classmethod
+    def watch(cls, watcher, flags):
+        data = cls._get()
+        iteration = data['iteration']
+        changed = bool(set(flags) & set(data['changes']))
+        return iteration == 0 or changed
+
+    @classmethod
+    def change(cls, flag):
+        data = cls._get()
+        data['pending'].append(flag)
+        cls._set(data)
+
+    @classmethod
+    def commit(cls):
+        data = cls._get()
+        data['changes'] = data['pending']
+        data['pending'] = []
+        cls._set(data)
 
 
 def dispatch():
