+++ docs/api.rst
@@ -4,11 +4,10 @@ Reactive API Documentation
 
 .. toctree::
 
+    charms.reactive.flags
     charms.reactive.decorators
     charms.reactive.helpers
     charms.reactive.relations
-    charms.reactive.bus
-    charms.reactive.cli
 
 .. automodule:: charms.reactive
     :members:
