+++ docs/charms.reactive.flags.rst
@@ -1,0 +1,16 @@
+charms.reactive.flags
+=====================
+
+.. rubric:: Summary
+
+.. automembersummary::
+    :nosignatures:
+
+    ~charms.reactive.flags
+
+.. rubric:: Reference
+
+.. automodule:: charms.reactive.flags
+    :members:
+    :undoc-members:
+    :show-inheritance:
